# leadforge-ai/backend/openapi.yaml
openapi: 3.0.3
info:
  title: LeadForge AI API
  description: |
    Complete API documentation for LeadForge AI - an automation platform that finds/tracks leads, 
    drafts and sends outreach, chats with prospects, extracts requirements, and generates/deploys websites.
    
    ## Authentication
    Currently no authentication required for local development.
    Production deployments should add JWT or API key authentication.
    
    ## Base URL
    - Development: `http://localhost:1100`
    - Production: Your deployed backend URL
    
  version: 1.0.0
  contact:
    name: LeadForge AI Support
    email: support@leadforge.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:1100
    description: Local development server
  - url: https://api.leadforge.ai
    description: Production server (if deployed)
...
paths:
...
  /track/open/{messageId}:
    get:
      tags:
        - Tracking
      summary: Track email open
      description: Tracking pixel endpoint (returns 1x1 transparent GIF)
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tracking pixel
          content:
            image/gif:
              schema:
                type: string
                format: binary
...
  /track/stats/{messageId}:
    get:
      tags:
        - Tracking
      summary: Get tracking statistics
      description: Retrieve open and click statistics for an email
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  opens:
                    type: integer
                  clicks:
                    type: integer
                  events:
                    type: array
                    items:
                      type: object
...